added LikeRepository doc and basic functionality

+++ src/main/java/com/makersacademy/acebook/repository/LikeRepository.java
@@ -1,9 +1,8 @@
 package com.makersacademy.acebook.repository;
 
 import com.makersacademy.acebook.model.Like;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LikeRepository extends CrudRepository<Like, Integer> {
+public interface LikeRepository extends CrudRepository<Like, Long> {
 
 }
